Changed f64 to f32

+++ src/controllers/time.rs
@@ -1,46 +1,46 @@
-use std::f64;
+use std::f32;
 use rand::{self, ThreadRng};
 
 use Resources;
 use super::Actions;
 use game_state::GameState;
-use geometry::{Advance, Position, Point};
-use models::{Bullet, Powerup, PowerupKind, Enemy, Particle, Vector};
+use geometry::{Advance, Position, Point, Vector};
+use models::{Bullet, Powerup, PowerupKind, Enemy, Particle};
 use util;
 
 // Constants related to time
-const BULLETS_PER_SECOND: f64 = 30.0;
-const BULLET_RATE: f64 = 1.0 / BULLETS_PER_SECOND;
+const BULLETS_PER_SECOND: f32 = 30.0;
+const BULLET_RATE: f32 = 1.0 / BULLETS_PER_SECOND;
 
-const ENEMY_SPAWNS_PER_SECOND: f64 = 1.0;
-const ENEMY_SPAWN_RATE: f64 = 1.0 / ENEMY_SPAWNS_PER_SECOND;
+const ENEMY_SPAWNS_PER_SECOND: f32 = 1.0;
+const ENEMY_SPAWN_RATE: f32 = 1.0 / ENEMY_SPAWNS_PER_SECOND;
 
-const TRAIL_PARTICLES_PER_SECOND: f64 = 20.0;
-const TRAIL_PARTICLE_RATE: f64 = 1.0 / TRAIL_PARTICLES_PER_SECOND;
+const TRAIL_PARTICLES_PER_SECOND: f32 = 20.0;
+const TRAIL_PARTICLE_RATE: f32 = 1.0 / TRAIL_PARTICLES_PER_SECOND;
 
-const POWERUP_SPAWNS_PER_SECOND: f64 = 1.0 / 30.0; // every ~30 seconds
-const POWERUP_SPAWN_RATE: f64 = 1.0 / POWERUP_SPAWNS_PER_SECOND;
+const POWERUP_SPAWNS_PER_SECOND: f32 = 1.0 / 30.0; // every ~30 seconds
+const POWERUP_SPAWN_RATE: f32 = 1.0 / POWERUP_SPAWNS_PER_SECOND;
 
 // Constants related to movement
 // Speed is measured in pixels per second
 // Rotation speed is measured in radians per second
-const ADVANCE_SPEED: f64 = 200.0;
-const BULLET_SPEED: f64 = 500.0;
-const ENEMY_SPEED: f64 = 100.0;
-const ROTATE_SPEED: f64 = 2.0 * f64::consts::PI;
-const STAR_BASE_SPEED: f64 = 50.0;
+const ADVANCE_SPEED: f32 = 200.0;
+const BULLET_SPEED: f32 = 500.0;
+const ENEMY_SPEED: f32 = 100.0;
+const ROTATE_SPEED: f32 = 2.0 * f32::consts::PI;
+const STAR_BASE_SPEED: f32 = 50.0;
 
-pub const PLAYER_GRACE_AREA: f64 = 200.0;
+pub const PLAYER_GRACE_AREA: f32 = 200.0;
 
 /// Timers to handle creation of bullets, enemies and particles
 pub struct TimeController {
     /// A random number generator
     rng: ThreadRng,
-    current_time: f64,
-    last_tail_particle: f64,
-    last_shoot: f64,
-    last_spawned_enemy: f64,
-    last_spawned_powerup: f64
+    current_time: f32,
+    last_tail_particle: f32,
+    last_shoot: f32,
+    last_spawned_enemy: f32,
+    last_spawned_powerup: f32
 }
 
 impl TimeController {
@@ -67,7 +67,7 @@ impl TimeController {
     /// Updates the game
     ///
     /// `dt` is the amount of seconds that have passed since the last update
-    pub fn update_seconds(&mut self, dt: f64, actions: &Actions, state: &mut GameState, resources: &Resources) {
+    pub fn update_seconds(&mut self, dt: f32, actions: &Actions, state: &mut GameState, resources: &Resources) {
         // You can run `cargo run --release --features "debug"` in order to run the game in
         // slow motion (assists in debugging rendering)
         #[cfg(feature = "debug")]
@@ -92,7 +92,7 @@ impl TimeController {
     }
 
     // Updates the position and rotation of the player
-    fn update_player(&mut self, dt: f64, actions: &Actions, state: &mut GameState) {
+    fn update_player(&mut self, dt: f32, actions: &Actions, state: &mut GameState) {
         if !state.world.player.is_dead {
             if actions.rotate_left {
                 *state.world.player.direction_mut() += -ROTATE_SPEED * dt;
@@ -107,7 +107,7 @@ impl TimeController {
     }
 
     // Adds, removes and updates the positions of bullets on screen
-    fn update_bullets(&mut self, dt: f64, actions: &Actions, state: &mut GameState, resources: &Resources) {
+    fn update_bullets(&mut self, dt: f32, actions: &Actions, state: &mut GameState, resources: &Resources) {
         // Add bullets
         if !state.world.player.is_dead && actions.shoot && self.current_time - self.last_shoot > BULLET_RATE {
             self.last_shoot = self.current_time;
@@ -118,9 +118,9 @@ impl TimeController {
                     let pos = state.world.player.front();
                     let dir = state.world.player.direction();
                     state.world.bullets.extend_from_slice(&[
-                        Bullet::new(Vector::new(pos, dir - f64::consts::PI / 6.0)),
+                        Bullet::new(Vector::new(pos, dir - f32::consts::PI / 6.0)),
                         Bullet::new(Vector::new(pos, dir)),
-                        Bullet::new(Vector::new(pos, dir + f64::consts::PI / 6.0)),
+                        Bullet::new(Vector::new(pos, dir + f32::consts::PI / 6.0)),
                     ]);
                 }
                 // If there was no powerup, shoot normally
@@ -143,7 +143,7 @@ impl TimeController {
         util::fast_retain(&mut state.world.bullets, |b| size.contains(b.position()));
     }
 
-    fn update_powerups(&mut self, dt: f64, state: &mut GameState) {
+    fn update_powerups(&mut self, dt: f32, state: &mut GameState) {
         for powerup in &mut state.world.powerups {
             powerup.update(dt);
         }
@@ -159,7 +159,7 @@ impl TimeController {
     }
 
     // Updates or removes particles on screen, adds particles behind player
-    fn update_particles(&mut self, dt: f64, state: &mut GameState) {
+    fn update_particles(&mut self, dt: f32, state: &mut GameState) {
         for particle in &mut state.world.particles {
             particle.update(dt);
         }
@@ -175,7 +175,7 @@ impl TimeController {
     }
 
     // Updates positions of enemies, and spawns new ones when necessary
-    fn update_enemies(&mut self, dt: f64, state: &mut GameState, resources: &Resources, time_slow: bool) {
+    fn update_enemies(&mut self, dt: f32, state: &mut GameState, resources: &Resources, time_slow: bool) {
         // Spawn enemies at random locations
         if self.current_time - self.last_spawned_enemy > ENEMY_SPAWN_RATE {
             self.last_spawned_enemy = self.current_time;
@@ -226,7 +226,7 @@ impl TimeController {
     }
 
     // Advance stars, wrapping them around the view
-    fn update_stars(&mut self, dt: f64, state: &mut GameState, time_slow: bool) {
+    fn update_stars(&mut self, dt: f32, state: &mut GameState, time_slow: bool) {
         for star in &mut state.world.stars {
             let speed = star.speed;
             let base_speed = if time_slow { 20.0 } else { STAR_BASE_SPEED };
