+++ src/geometry/size.rs
@@ -5,13 +5,13 @@ use super::Point;
 /// A `Size` represents a region in space
 #[derive(Clone, Copy, Default)]
 pub struct Size {
-    pub width: f64,
-    pub height: f64
+    pub width: f32,
+    pub height: f32
 }
 
 impl Size {
     /// Returns a new `Size` of the given dimensions
-    pub fn new(width: f64, height: f64) -> Size {
+    pub fn new(width: f32, height: f32) -> Size {
         Size { width: width, height: height }
     }
 
@@ -22,12 +22,12 @@ impl Size {
     }
 
     /// Returns a random x coordinate within the bounds of this `Size`
-    pub fn random_x<R: Rng>(&self, rng: &mut R) -> f64 {
+    pub fn random_x<R: Rng>(&self, rng: &mut R) -> f32 {
         rng.gen_range(0.0, self.width)
     }
 
     /// Returns a random y coordinate within the bounds of this `Size`
-    pub fn random_y<R: Rng>(&self, rng: &mut R) -> f64 {
+    pub fn random_y<R: Rng>(&self, rng: &mut R) -> f32 {
         rng.gen_range(0.0, self.height)
     }
 }
