+++ src/geometry/vector.rs
@@ -1,8 +1,7 @@
-use std::f64;
-
+use std::f32;
 use rand::Rng;
 
-use geometry::{Point, Size};
+use super::{Point, Size};
 
 /// A `Vector`
 #[derive(Clone, Default)]
@@ -10,23 +9,23 @@ pub struct Vector {
     /// The position of the vector
     pub position: Point,
     /// The direction angle, in radians
-    pub direction: f64
+    pub direction: f32
 }
 
 impl Vector {
     /// Returns a new `Vector`
-    pub fn new(position: Point, direction: f64) -> Vector {
+    pub fn new(position: Point, direction: f32) -> Vector {
         Vector { position: position, direction: direction }
     }
 
     /// Returns a random `Vector` within the given bounds
     pub fn random<R: Rng>(rng: &mut R, bounds: Size) -> Vector {
-        Vector::new(Point::random(rng, bounds), rng.gen::<f64>() * 2.0 * f64::consts::PI)
+        Vector::new(Point::random(rng, bounds), rng.gen::<f32>() * 2.0 * f32::consts::PI)
     }
 
     /// Consumes the vector and returns a new one with inverted direction
     pub fn invert(mut self) -> Vector {
-        self.direction -= f64::consts::PI;
+        self.direction -= f32::consts::PI;
         self
     }
 }
@@ -36,18 +35,18 @@ impl Vector {
 macro_rules! derive_position_direction {
     ($t:ty) => {
         impl ::geometry::Position for $t {
-            fn x(&self) -> f64 { self.vector.position.x }
-            fn y(&self) -> f64 { self.vector.position.y }
-            fn x_mut(&mut self) -> &mut f64 { &mut self.vector.position.x }
-            fn y_mut(&mut self) -> &mut f64 { &mut self.vector.position.y }
+            fn x(&self) -> f32 { self.vector.position.x }
+            fn y(&self) -> f32 { self.vector.position.y }
+            fn x_mut(&mut self) -> &mut f32 { &mut self.vector.position.x }
+            fn y_mut(&mut self) -> &mut f32 { &mut self.vector.position.y }
         }
 
         impl ::geometry::Advance for $t {
-            fn direction(&self) -> f64 {
+            fn direction(&self) -> f32 {
                 self.vector.direction
             }
 
-            fn direction_mut(&mut self) -> &mut f64 {
+            fn direction_mut(&mut self) -> &mut f32 {
                 &mut self.vector.direction
             }
         }
